+++ backend/packages/agents/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @cat-factory/agents
 
+## 0.110.3
+
+### Patch Changes
+
+- Updated dependencies [ee6601e]
+  - @cat-factory/contracts@0.233.0
+  - @cat-factory/kernel@0.234.1
+  - @cat-factory/prompt-fragments@0.15.59
+
+## 0.110.2
+
+### Patch Changes
+
+- Updated dependencies [937d4af]
+  - @cat-factory/contracts@0.232.0
+  - @cat-factory/kernel@0.234.0
+  - @cat-factory/prompt-fragments@0.15.58
+
+## 0.110.1
+
+### Patch Changes
+
+- Updated dependencies [2580fee]
+- Updated dependencies [eb4ca17]
+  - @cat-factory/kernel@0.233.0
+  - @cat-factory/contracts@0.231.0
+  - @cat-factory/prompt-fragments@0.15.57
+
 ## 0.110.0
 
 ### Minor Changes