+++ backend/packages/caching/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @cat-factory/caching
 
+## 0.14.12
+
+### Patch Changes
+
+- @cat-factory/kernel@0.234.1
+
+## 0.14.11
+
+### Patch Changes
+
+- Updated dependencies [937d4af]
+  - @cat-factory/kernel@0.234.0
+
+## 0.14.10
+
+### Patch Changes
+
+- Updated dependencies [2580fee]
+- Updated dependencies [eb4ca17]
+  - @cat-factory/kernel@0.233.0
+
 ## 0.14.9
 
 ### Patch Changes